Implement comparator
This change adds the functionality to penalize inferior architectures in order
to sort a given population. This is done by creating a temporary vector that
stores the essential information (architecture index, critical path, and area)
as a tuple for each architecture. The temporary vector is sorted, and the
index in the sorted vector is used as the penalty value (so the architectures
towards the front gets less penalty). These penalty values are added up to
either `speed_penalty' or `area_penalty', which are used finally to sort the
architecture population.