Use percentages for measuring performance improvement
Percentages are used due to the fact that when there are multiple benchmarks
and multiple parameters with different units, it is difficult to normalize
them. By comparing the current population against the performance of the first
generation, we get a decent metric of performance change. Since we want
architectures with smaller critical path and smaller area, the lower the
ration the better. This deprecates the "sort by how many an architecture wins
to its siblings" approach.

This change also fixes the logic error regarding selection. Prior to this
change, all of the current population was considered when performing crossover
and mutation. However, only selected entities should be allowed to perform
crossover and mutation. `num_selected' parameter is added to control how many
architectures should be considered in the process of selecting the pool for
crossover and mutation.